new-agent-system/requirements.txt
# Multi-Agent Booking Extraction System Requirements
# Core dependencies for the intelligent booking data extraction system

# OpenAI API for GPT-4o-mini agents
openai>=1.3.0

# AWS SDK for Textract OCR processing
boto3>=1.34.0
botocore>=1.34.0

# Data processing and manipulation
pandas>=2.0.0
numpy>=1.24.0

# Image processing for table screenshots
Pillow>=10.0.0

# PDF processing support
PyPDF2>=3.0.0

# JSON handling and validation
jsonschema>=4.19.0

# Logging and configuration
pyyaml>=6.0

# HTTP requests for API calls
requests>=2.31.0

# Type hints and validation
typing-extensions>=4.8.0

# Date and time parsing
python-dateutil>=2.8.0

# Regular expressions enhancements
regex>=2023.8.8

# Optional: Progress bars for batch processing
tqdm>=4.65.0

# Optional: Environment variable management
python-dotenv>=1.0.0

# Development and testing dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Streamlit for web interface
streamlit>=1.28.0

# Optional: Enhanced CSV processing
openpyxl>=3.1.0